Move unstructured to its own package under v1
It is a versioned type.
