+++ pkg/runtime/types.go
@@ -122,17 +122,6 @@ type Unknown struct {
 	ContentType string `protobuf:"bytes,4,opt,name=contentType"`
 }
 
-// Unstructured allows objects that do not have Golang structs registered to be manipulated
-// generically. This can be used to deal with the API objects from a plug-in. Unstructured
-// objects still have functioning TypeMeta features-- kind, version, etc.
-// TODO: Make this object have easy access to field based accessors and settors for
-// metadata and field mutatation.
-type Unstructured struct {
-	// Object is a JSON compatible map with string, float, int, []interface{}, or map[string]interface{}
-	// children.
-	Object map[string]interface{}
-}
-
 // VersionedObjects is used by Decoders to give callers a way to access all versions
 // of an object during the decoding process.
 type VersionedObjects struct {
